fix(gpu): correct Q6_K dequant offset in quant_matmul_q6_k_kernel
GPU kernel subtracted 32.0 instead of d*sc*32 for the Q6_K value
centering offset. Formula was d*sc*v6 - 32.0, should be d*sc*(v6-32).

This caused the SSM output projection to produce constant ~365 for all
2048 output elements when the gated-norm output had non-zero mean,
completely destroying the hidden state and producing garbage ('endah').

Fix: compute d*sc*(v6-32) as three separate double multiplications to
match the CPU dequant in gguf_reader.c dequantize_q6_K_row().

+++ src/gpu_quant_matmul.cu
@@ -111,19 +111,19 @@ __global__ void quant_matmul_q6_k_kernel(const float *x, const uint8_t *W_q,
                 uint8_t h = qh[l];
                 if (idx0 < b * QK_K + rem) {
                     int8_t v6 = (int8_t)((l0 & 0xF) | ((h >> 0) & 3) << 4);
-                    sum += (double)x[idx0] * ((double)d * sc[is+0] * v6 - 32.0);
+                    sum += (double)x[idx0] * (double)d * (double)sc[is+0] * (double)(v6 - 32);
                 }
                 if (idx1 < b * QK_K + rem) {
                     int8_t v6 = (int8_t)((l32 & 0xF) | ((h >> 2) & 3) << 4);
-                    sum += (double)x[idx1] * ((double)d * sc[is+2] * v6 - 32.0);
+                    sum += (double)x[idx1] * (double)d * (double)sc[is+2] * (double)(v6 - 32);
                 }
                 if (idx2 < b * QK_K + rem) {
                     int8_t v6 = (int8_t)((l0 >> 4) | ((h >> 4) & 3) << 4);
-                    sum += (double)x[idx2] * ((double)d * sc[is+4] * v6 - 32.0);
+                    sum += (double)x[idx2] * (double)d * (double)sc[is+4] * (double)(v6 - 32);
                 }
                 if (idx3 < b * QK_K + rem) {
                     int8_t v6 = (int8_t)((l32 >> 4) | ((h >> 6) & 3) << 4);
-                    sum += (double)x[idx3] * ((double)d * sc[is+6] * v6 - 32.0);
+                    sum += (double)x[idx3] * (double)d * (double)sc[is+6] * (double)(v6 - 32);
                 }
             }
             ql += 64; qh += 32; sc += 8;
